+++ example/src/main/java/com/getbase/android/db/example/ui/WrapperTransformationListFragment.java
@@ -9,11 +9,12 @@ import com.google.common.base.Function;
 import android.content.Context;
 import android.database.Cursor;
 import android.os.Bundle;
-import android.support.v4.app.ListFragment;
-import android.support.v4.content.Loader;
 import android.view.View;
 
 import java.util.List;
+
+import androidx.fragment.app.ListFragment;
+import androidx.loader.content.Loader;
 
 public class WrapperTransformationListFragment extends ListFragment implements LoaderHelper.LoaderDataCallbacks<SectionedPeopleList> {
 
@@ -48,7 +49,6 @@ public class WrapperTransformationListFragment extends ListFragment implements L
               return new Person(firstName, secondName);
             }
           })
-          .lazy()
           .transform(new Function<List<Person>, SectionedPeopleList>() {
             @Override
             public SectionedPeopleList apply(List<Person> people) {