+++ library/src/androidTest/java/com/getbase/android/db/test/TestActivity.java
@@ -1,0 +1,13 @@
+package com.getbase.android.db.test;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.FragmentActivity;
+
+public class TestActivity extends FragmentActivity {
+
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+  }
+}