+++ library/src/main/java/com/getbase/android/db/loaders/CancellableFunction.java
@@ -1,0 +1,7 @@
+package com.getbase.android.db.loaders;
+
+import androidx.core.os.CancellationSignal;
+
+public interface CancellableFunction<TIn, TOut> {
+  TOut apply(TIn input, CancellationSignal signal);
+}