+++ changelog.md
@@ -1,6 +1,33 @@
 Change Log
 ==========
 
+Version 0.15.0 *(2021-01-15)*
+----------------------------
+
+ * Migration to AndroidX
+ * Fix ConcurrentModificationException
+
+Version 0.14.0 *(2017-08-28)*
+----------------------------
+
+ * Added support for loader cancellation in `ComposedCursorLoader`.
+ * Dropped support for lazy row transformation.
+
+Version 0.13.4 *(2017-04-03)*
+----------------------------
+
+ * Fix issue with ComposedCursorLoader and boxed primitives..
+
+Version 0.13.3 *(2016-11-22)*
+----------------------------
+
+ * Provide more info when exception is thrown from `AbstractLoader.onNewDataDelivered()`
+
+Version 0.13.2 *(2016-05-17)*
+----------------------------
+
+ * Provide more info when exception is thrown from `ComposedCursorLoader.loadInBackground()`
+
 Version 0.13.1 *(2015-12-05)*
 ----------------------------